api/tests: Introduce a test mock
Add a mock implementation of virtcontainers via a new vcMock package
that defines a VCMock type, along with an interface that all
virtcontainers implementations are required to implement (VC).

Since the Pod and Container fields are entirely private, it is also
necessary to introduce interfaces for both those types (VCPod and
VCContainer respectively) to allow the mock implenentation to be fully
functional.

As some of the virtcontainers API calls return Pod and Container types,
the API has also had to be modified slightly to return the interface
type (required by the go language). This change then further required
that any functions that need to manipulate the real types (Pod and
Container for example) have to cast the interface types back to the
underlying types using type assertions.

Note that this is potentially an API-breaking change: if clients have
specified an explicit type, the code will need to be updated to change
the explicit types (Pod, Container) to the interface types (VCPod,
VCContainer).

For example, the following code:

    var p vc.Pod
    p = CreatePod(config)

... will need to be changed to either:

    var p vc.VCPod
    p = CreatePod(config)

... or more simply:

    p := CreatePod(config)

As a convenience, a new VCImpl type is introduced that also implements
the VC interface along with a test to ensure that both provided
implementations are usable by consumers (TestVCImplementations).

Fixes #346.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>