qemu: Enable specific flags if running on VMM
In case our Qemu process runs on a VMM, that is as a nested VM, we
want to add flags "disable-modern=true" and "pmu=off" in order to
disable virtio 1.0 and the power management unit.
That's the only way we have found to prevent our Clear Containers
VM from hanging because of the KVM known bugs.

Notice this patch enables "disable-modern=true" only if the device
is driven by a virtio driver.

Fixes #342

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>