Realtime & Mlock: Enable support for realtime and memory locking
Add support for Realtime and Mlock support. This allows enabling
of realtime for QEMU. Additonally it allows the VM memory to be
swapped out when Realtime is enabled.

Realtime can be enabled for latency sensitive workloads.
A combination of prealloc, realtime and mlock gives the lowest
latency VM.

Enabling Realtime with Mlock=false allows for higher density at
the cost of higher latency.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>