Nesting: Disable nesting checks
This is used to disable the customizations performed when
we are running on top of another VM. This can be used when
running on top of KVM to obtain the best performance.

This check should not be disabled when running on hypervisors
other than KVM. Doing so will result in failure to launch.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>