pod: VMs: more logging, better timeouts
VMs can take a little while to get going depending on the
circumstances - such as a very busy or a large (cores and RAM)
VM - 1s is not long enough to wait in such cases.

To aid in such cases:
- Bump the startup timeout (it is better to wait 'too long' to realise
  something broke, rather than firing false failures)
- Add some more logging to make it clearer what happened

For now we make the concious decision to leave the VM shutdown
timeout at 1s.

Fixes: #390

Signed-off-by: Graham Whaley <graham.whaley@intel.com>