filesystem: Custom marshal and unmarshal device interface objects
The idea is to move []Devices from ContainerConfig to
Container and store them in a separate devices.json file.
This moves the custom json handling at ContainerConfig level
to Device level simplifying things and is inline with our overall
design approach.

Introduce a TypeDevice struct which will act as an intermediate
representation for marshalling and unmarshalling Device interface
implementations.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>