network: Handle marshalling/unmarshalling of Endpoint interface array
Since NetworkNamespace now contains an interface array,
we need to add custom marshalling and unmarshalling methods for
NetworkNamespace to handle the interface objects correctly.

We introduce a new struct TypedJSONEndpoint to store the type of
the object as a separate json attribute and the endpoint itself
as a json.RawMessage. We first extract the type from this struct
while unmarshalling, so that the actual endpoint can be
unmarshalled into the right Endpoint struct depending on the type
that was marshalled first.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>