lock: Use shared locking for read only cases
In order to improve the performance of the locking and reduce the
bottlenecks, this commit changes the lock type for functions only
reading data from the storage, but never writing anything.

Also, it handles the special case where, while reading through the
different container states, it realizes that the shim is not running.
In that case, we want to use an exclusive lock because it is going to
update the container state.

Fixes #421

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>