devices: Change check for block devices.
The earlier check for block devices was restrictive, just check
device type instead. This is to support cases such as when a block
device is loop mounted. We will need to explicitly check for
the path matches, when we add support for other vitio drivers such
as virtio-scsi for passing block devices to the VM.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>