devices: Ignore error from missing sysfs path under /sys/dev
Ability to obtain device names from major-minor number of device
was added in kernel 2.6.27 version:
https://github.com/torvalds/linux/commit/e105b8bfc769b0545b6f0f395179d1e43cbee822

But for certain devices like /dev/cuse, /dev/fuse, this sysfs
interface may not exist. This may also depend on the kernel
version used.

We are primarily using this sysfs interface to obtain device name in
case the renamed device name was passed to the runtime.

To accomodate these devices, which are passed by docker by default
for priveleged/non-priveleged containers, we ignore the missing file
error for /sys/dev files and assume device name passed in the oci spec
file for such devices is not the renamed device name.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>