.ci: Handle cases where backward compatibility is broken
When a pull request is submitted to virtcontainers, the CI will run,
testing the whole compatibility with Clear Containers runtime.
Unfortunately, when this PR involves some changes in virtcontainers
API or structures, the CI fails because the runtime cannot be built.

This commit introduces the possibility to specify a new field in the
commit message of the last commit of the pull request submitted to
virtcontainers.

The documentation has been updated to explain how to use this tag.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>