pkg: qemu: Move qemu CLI wrapper from ciao
Move the qemu CLI wrapper from ciao repository to our own package
in virtcontainers. This way, no need to revendor ciao every time
we make a change to this package. The other reason is that ciao
is not really using all the things that we have recently introduced
and it will simplify our process.

Fixes #431

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>