network: Replace CNI Result structure with a larger one
The goal of this commit is to replace the dependency on CNI structure
types.Result with a new one. The new structure relies on very generic
Go structures such as net.Interface, netlink.Route, and net.IPNet in
order to provide a lot of information about the underlying network.

The choice has been made to translate only the DNS information coming
from the CNI network implementation because other data can be found
through a simple network scan. And because this scan has to be
performed anyway (need MTU which is not provided by CNI result for
instance), it is easier to fill network interfaces, IP addresses and
routes structures from scratch.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>