network: cni: Fix lack of support for SRIOV
CNM has been previously covered so that it can create some physical
endpoints, but CNI was not taking them into account. This was due to
the fact CNI was not scanning the network, but was relying on CNI
results only.

This commit adds the support for physical devices the same way it's
done with CNM. A scan is performed and the devices types are deduced
from there. Notice that we still want to keep DNS info from CNI, and
that's the reason why we don't override the DNS with what the scan
returned since a scan cannot provide DNS info.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>