network: Add interface type to our network info
In order to store every information we get from the netlink Link
interface, this commit changes our NetworkInfo structure so that
it now includes the Link type.

This will prevent us from doing further scans in order to retrieve
the type. And note that we are solving a potential issue for physical
devices since a later scan would not find the link information.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>