storage: Simplify fetches from storage
The way we have been handling the fetch was too rigid and was forcing
us in doing several changes when adding new types of files.

With this commit, we rely on the magic of Go unmarshalling. The JSON
package handle all the complexity for us, allowing much more flexible
code.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>