api: Simplify drive hotplugging
The way CreatePod() was implemented, it was making an extra call
to hotplug all the drives. But when StartPod() is called, it leads
eventually to Container.start() to be invoked, resulting into a
hotplug of the drives related to this container.

A mechanism based on a flag was in place to detect if the drive had
been previously hotplugged, but there is no need for such thing
anymore. Indeed, we know that we are hotplugging only from the
container implementation.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>