pod: Improve pod create performances
In order to optimize pod creation, this patch does not wait for the
VM after it has been started. Instead, it will wait for it during the
"start" stage. This will allow our VM to be started while the caller
could potentially do other things between create and start. Globally,
this improves the boot time of our containers.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>