qemu: Provide unique QMP sockets names
Because the previous commit introduced some improvements by not
waiting for the end of the VM, it also introduced a side effect
issue. When a pod is deleted, the caller can consider this pod
ID as not used anymore, meaning it can start a new pod with the
same ID. The problem being the new QMP sockets will be created,
but if the previous VM terminates after this action, it will
remove the freshly created sockets because they will have the
exact same names in the exact same directory.

The way to solve this issue is to generate a unique UUID that we
can concatenate with the socket name to make it unique. There are
some requirements to meet about the QMP socket path length this
patch is also handling.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>