qemu: Fix disconnectCh issue and simplify it
The fact that waitPod() implementation for QEMU tries to connect
to the VM from a loop, means that every time QMPStart() returns,
the disconnect channel is closed. We need a new channel to be setup
every time we call QMPStart() again.

Also, this commit takes care of removing disconnectCh field from
qmpChannel structure since this is not used.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>