cni: Fix bug while updating endpoints after scanning net namespace
After scanning the network namespace, we need to update the scanned
endpoints with the DNS information, and use the scanned endpoints
thereafter discarding the endpoints structures we had created prior
to calling into the CNI plugin.

There was a bug in the update, which was causing the entire endpoint
structure being assigned back to the scanned endpoints causing us to
lose information from the scan including if a virtual or physical
endpoint was found in the scan.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>