vhost-user enabling
During network namespace scan, we check for a "dummy interface" which
is created to indicate a vhost-user socket is available for consumption.

The interface is identified and vhost-user socket is verified by
searching for /tmp/vhostuser_<ipv4 addr>/vhu.sock, where the ipv4
address matches the dummy interface created in the network namespace.

Generic vhost-user enabling for non-network devices (like SCSI) still
needs to be tested.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>