vhost-user: enabling for vhost-user network devices
During network namespace scan, we check for a "dummy interface" which
is created to indicate a vhost-user socket is available for consumption.

The interface is identified and vhost-user socket is verified by
searching for /tmp/vhostuser_<ipv4 addr>/vhu.sock, where the ipv4
address matches the dummy interface created in the network namespace.

Block and SCSI vhost-user device types are added as a placeholder, but
generic device support for these is not yet implemented.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>