device: make a more generic function for hypervisor args
Before we had a block device specific function which would create a
drive-(id) string which could be used to uniquely identify a drive on a
hypervisor commandline

Made this more generic so you could do the same for other device types,
such as -net or -scsi, etc.  Updated existing calls to make use of
generic function and moved it from device to hypervisor.go.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>