network: refactoring and cleanup of CNI path
Some of the function naming and process around CNI's add() were overly
complicated and/or obfuscated.

Initially, the CNI "add" function would create a virtual device which
would then be used to store the results obtained from calling the CNI
plugin.  This patch removes that, and instead keeps the netinfo results
for application to the endpoint device when we identify it post-scan.

Renamed addVirtInterfaces and deleteVirtInterfaces functions.
add is now invokePlugins, since this is what is carried out, and
the "virt" is removed from delete since there is nothing specific
to virtual in this case (it just invokes delete command to the plugin).
The function prototype for invokePlugins was also modified to return
the resulting netinfo from the CNI call, allowing the DNS information
to be used when updating the endpoints after the network namespace
is scanned.

The logic has been updated to reflect that a pod will only have a single
network namespace.  This simplified the code in a couple of locations,
since we know the initial numInterfaces is always '1' - no need to loop.
This single interfaces is described as 'CniPrimaryInterface' in the
sources.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>
Signed-off-by: Manohar Castelino  <manohar.r.castelino@intel.com>