agent: Introduce new functions to agent interface
We need to introduce two new functions to the agent interface in
order to be able to handle different proxy cases.

On one hand, we need to be able to set a proxy URL from the proxy
implementation, allowing the agent implementation to transparently
connect to the provided URL. The agent implementation won't know if
this URL is a real proxy, or if it is the direct URL of the VM. But
either way, it will connect the same way, and this information has to
be set by the only piece who has the actual knowledge, that is proxy
implementation.

On the other hand, we need to provide a way to retrieve properly the
URL setup by the agent implementation and exposed by the VM. This is
needed by the proxy implementation to be able to connect the VM, but
this can also be useful in case of no_proxy implementation, when we
actually need to provide the agent with the VM URL so that it can
directly connect to it.

Fixes #524

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>