container: Generate process token when not set
Some proxy (e.g. Kata Containers proxy) will not generate
a process token, they expect that the runtime will generate
one and pass it to the agent when either creating a container
or exec'ing a process.

So here we generate a process token when the specified proxy
does not provide one.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>