pod: Add important comment for Cmd type
The `Cmd` type encodes the user and group details that the workload will
be run as.

It currently encodes those values as strings, although since the OCI
spec and the Kata gRPC protocol encode such values as integers, we would
ideally like to change `Cmd` so that the user and group details are also
encoded as integer values (with a conversion to strings for sending such
details to hyperstart which uses string types for them).

Add a comment explaining why we cannot change the type of the user and
group fields to an integer type.

Fixes #533.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>