ci: Handle complex revendoring cases
This commit relies fully on dep tool so as to vendor the latest code
coming from a pull request on virtcontainers repository.

The information about the github repo URL and the commit revision are
both passed from Jenkins (as environment variables), and the script
relies on those values in order to override Gopkg.toml for both proxy
and runtime repositories.

The last step is to run dep tool, leaving the revendoring complexity
(including nested vendoring sometimes) being handled by dep.

Fixes #513

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>