shim: Add a Kata shim mock implementation
It's different from the cc-shim one as it uses different parameters.
In order to not duplicate too much code between the 2 mock shims, we
factorize the main routine in mock.go. The assumption here is that
any virtcontainers shim will need an agent URL, a process token and a
container ID as its parameters.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>