qemu: refactor/simplify addDevice function
In adding vhost-user support, a type switch was added.  For consistency
decided to rewrite the rest of the function to also use a type switch.

As a next step we should consider updating the function parameters for
addDevice and equivalent functions in the hypervisor interface since the
deviceType is not necessary.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>