qemu: adjust QMP naming to avoid non-unique truncation
The full name of the QMP monitor and control sockets was well
beyond the 107 character limit, resulting in the full name to
be truncated during creation.  As a result, only a single socket
was created.

Instead of creating a socket at
<parentDirPath>/<q.state.UUID>-<socketName>, create it at
<parentDirPath>/<socketName>-<q.state.UUID>, keeping as much of the
unique socket name as feasible.

Also, to limit truncation, reduced size of the monitor and control
socket names.

Fixes #552

Signed-off-by: Eric Ernst <eric.ernst@intel.com>