pkg/oci: Clarify resource calculation comment
We use a couple of relativley simple formulas to work out how much
RAM and how many vCPUs to allocate to the VM based on the
Resource request we are passed in.
Improve the comment on these calculations to make it clear which
way they round and if they impose a floor to the resulting value.

Fixes: #554

Signed-off-by: Graham Whaley <graham.whaley@intel.com>