container: Create container before starting it
Although hyperstart does not know about creating a container
(it does create+start from NEWCONTAINER), Kata does.
As a consequence we need to call the agent createContainer()
method before the startContainer() one. We make that call from
container.start() because from there we know for sure that
the agent is up and running already.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>