kata_agent: Replace OCI mounts source paths
With Kata, all OCI mounts are done by the agent. But
we need to modify the mounts source as they're no longer
are relative to the host but to the guest instead.
So we bind mount the host mounts into the guest shared
mount point and replace the OCI mounts source with the
guest one.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>