scsi: Add support for scsi driver for hotplugging drives
Introduce a configuration option for using either virtio-scsi
or virtio-block for hotplugging drives with virtio-scsi being
the default driver. Hotplug block devices based on this
confuration.
In order to hotplug SCSI drives, we need to coldplug a SCSI
controller while starting up qemu.

With virtio-scsi, use the index at which the drive is attached
to come up with a SCSI id and LUN. This will be used by the agent
to detect the drive name.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>