logging: Set source field for oci and hyperstart packages
Change the `SetLogger()` API calls in the `oci` and `hyperstart`
packages so that they set the `source` logger field. This allows
callers making use of these packages to see the true source of the log
messages (previously, it would appear that log calls made by these two
packages would be coming *from* the caller itself).

Fixes #571.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>