pod: Wait for the VM and start the pod in startVM
We need to start the pod from startVM as some architecture (namely Kata)
needs to create containers before startPod(). Which means the pod inside
the VM needs to exist before startPod() is called in order to actually
create containers linked to it. For that reason we need to start the pod
(and thus wait for the agent signal) from startVM, i.e. from the
createPod() step.

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>