vendor: Update kata agent vendoring
We need this revendoring to fix the spec converter. We are relying
on the OCI to GRPC spec converter provided by kata agent, and this
has been recently fixed. We need this fix to go in for virtcontainers.

Shortlog:

045c2cf grpc: Split execProcess in more readable functions
71dd0b8 protocols: grpc: Fix the spec converter regarding nil pointers
532325c grpc: Move sequencing from StartContainer to CreateContainer
3dd881f grpc: Reference appropriate process when starting a container
b2f30e1 grpc: Lock the container creation
92a2a6a grpc: Remove unused code parts
d3a27ca agent: Add missing nsenter import
4f73ac9 mount: Create destination directory for 9pfs as well
1327404 mount: Do not shadow absSource
0759700 grpc: Return nil from SignalProcess() if container already stopped
9d6298b vendor: Re-vendor libcontainer for SIGKILL fix
3d44912 agent: allow exec before start
35e36c6 test: Increase unit test coverage

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>