proxy: Rework proxy interface
The proxy interface was completely Clear Containers specific. That's
the reason why it needed to be changed so that virtcontainers could
meet its primary goal, that is supporting several implementations of
agents.

This commit reworks the proxy interface by narrowing the set of
functions to two simple ones: start() and stop().

The whole legacy interface is handled directly from the hyperstart
agent implementation, making this tied to the CC proxy implementation.
This is not very different from our previous state where we were also
tied between cc agent and cc proxy, but the huge benefit is about
other implementations like Kata where we actually have something
generic. We can use kata proxy or no proxy seamlessly and it would
work the same way, making the proxy a sub-layer of the agent.

Fixes #518

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>