agent: Cleanup agent initialization
This commit simplifies and cleans up agent implementations by removing
unused volumes and sockets.

It re-organizes the way the initialization is done, that is by relying
on the agent configuration in order to fill the agent object properly.
Indeed, the agent was saving the modified config, which is not an
acceptable thing to do.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>