agent: Store URL and ProxyPid info from each implementation
This commit creates a new file agent.json in order to store specific
data such as the URL the agent implementation could need to interact
with or the proxy PID retrieved after the agent started the proxy.

Those data can vary from one implementation to another, but the goal
is to be able to store anything the agent needs to fetch, instead of
having those info inside the Pod structure which does not make a lot
of sense.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>