qemu: Log default kernel params
Add a debug-only log call that logs the default kernel parameters. This is
useful for users to identify which parameters are being set by virtcontainers,
which are set by the runtime and which are user-specified.

A new `defaultKernelParams()` function was added to return these default
parameters. This allow the default kernel parameters to be logged once prior
to hypervisor startup and before they are combined with the defaults passed by
the runtime.

Fixes #585.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>