container: Replace fetchContainer() with findContainer()
The function fetchContainer() defined in container.go was
always called from the high level API, in api.go, after a
call to fetchPod() defined in pod.go. Because fetchPod()
rebuilds everything from scratch, based on the pod and
containers configurations, there is no need for a function
fetching containers from the storage. What we really need
is a function returning a container pointer based on the list
of existing containers pointers. This function is the new function
introduced by this commit, called findContainer().

Fixes #588

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>