api: Add new container pointer to pod structure
This commit makes sure after we have created a new container for the
first time, this container pointer should be added to the containers
list held by the pod structure.

This way, we ensure the containers list held by the pod structure is
up to date, and that we don't have container pointer duplicate.

Fixes #588

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>