container: Move devices and mounts creation to createContainer()
Instead of waiting for the start a container to get its mounts and
devices processed, this commit moves those actions earlier in the
sequence, that is to CreateContainer().

The reason behind this is that in case of Kata Containers, we expect
all resources to be created so that we can actually create the
container from the agent, when CreateContainer() is called.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>