mount: Explicitely modify and store container mount list
The way container mount list was modified by each agent implementation
was very implicit and not obvious. In order to get a clearer code and
avoid some c.storeMounts() from container.go, not knowing if the mount
list was actually modified, this patch makes explicit from each agent
implementation the need to modify and store the container mount list.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>