qemu: add support for ARM
This patch introduces a new interface called qemuArch that contains
the functions needed to create qemu specific architecture implementations.
This patch also includes three qemu specific architecture
implementations: qemuArchBase, qemuAmd64 and qemuArm64, since all architectures
use similar qemu command lines, the base structure is qemuArchBase that is an
agnostic implementation to start virtual machines, qemuAmd64 and qemuArm64
inherit and override (if needed) the functions of qemuArchBase, therefore architecture
specific optimizations are in qemuAmd64 and qemuArm64.
For example the `appendImage` function defined by qemuArchBase uses a block
device to provide the image to the VM, qemuArm64 inherits and doesn't
overwrite it, but qemuAmd64 inherits and overwites it to provide the
image throught NVDIMM and DAX.

fixes #608
fixes #610
fixes #611
fixes #612

Signed-off-by: Julio Montes <julio.montes@intel.com>