kata_agent: Empty namespaces paths
As a specification constraint, this commit ensures that every
namespace, excepted for network namespace which is completely
removed, we should empty the path from the agent implementation.

Indeed, the namespaces paths provided cannot be found by the agent
running inside the VM, and that's why they have to be re-created
from scratch.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>