kata_agent: Ignore resources from the specification
This commit is a temporary way to have Kata Containers working with
CRI-O. Indeed, the Kata agent does not support yet the way to apply
properly cgroups constraints and this patch make sure no constraints
are passed to the agent.

This has to be removed as soon as those constraints can be handled
from the agent.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>