pod: Delete containers when deleting the pod
Instead of deleting only the pod from a DeletePod(), this commit
also deletes the associated containers. This is done by relying
on the delete() function defined from container.go.

As a consequence of calling delete() from container.go, we make
sure the shims are stopped for every container. This leads to the
removal of stopShims() which is not needed anymore.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>