api: Stop the VM when stopping the pod
The pod being very tied to the VM itself, this commit tries to make
the API fitting better this definition by moving the destruction of
the VM when the pod is stopped, instead of when the pod is being
deleted. Following the same idea, the network and the hooks should
be removed after the pod has been stopped, leaving only the removal
of internal structures and resources created into the filesystem
being done by the call to DeletePod().

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>