pkg: nsenter: Introduce a generic nsenter package
This package goal is to make very simple the execution of any code
inside any set of namespaces. Based on the CNI implementation of the
package 'ns' on the plugins repository, this new package takes a list
of namespaces as input and a routine that should be called once all
those namespaces have been entered.

This way, this package provide a generic way to spawn any binary into
the expected set of namespaces. It exports only one function NsEnter()
making the usage very simple.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>