shim: Add ability to enter any set of namespaces
This new commit adds to the shim the ability to be run within any
existing namespace. This will be very useful in order to start shims
representing exec'ed processes related to an existing container. In
this case, we will need to start those shims in the namespaces of the
container process shim.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>