shim: Start shims inside PID namespace
The shim representing the container process needs to be in its own
PID namespace as the "init" process. The reason is that we need to
make sure it will wait for its exec'ed processes in the same PID
namespace to be killed before to return. This happens inside each
agent, but if this is not reproduced on the host, we get some race
conditions leading the shim container process to return before its
exec'ed processes are killed. And this leads those shims to never
return their exit codes.

This commit solves this issue by making sure a new shim representing
a container process enters a new PID namespace, and any exec'ed
process related to this container enters this PID namespace previously
created.

Fixes #613

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>