kata_agent: Wait for rootfs if it is a hotplugged device
In case the rootfs is not a directory bind mounted through 9pfs, but
instead is a block device hotplugged into the VM, we need a way to
make sure the agent will wait for this device to properly show up
before to process further.

Relying on the property exposed by the Device structure of the Kata
agent protocol, we provide an empty ContainerPath for our device,
meaning it will be waited for, but the agent will not try to update
it through the OCI spec device list.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>