docs: Developer: must develop/test in main repo
You cannot build and test in a repo fork, as the sub-pkg
references do not move with the fork, and we fail due to
type difference failures.

Add details on how to add your user fork as a remote to the
main repo so you can develop/PR.

Also update details on installing required components into
/tmp, and the inherent weakness of that.

Note you need 'CI=true' to usefully run `make check`.

Fixes: #642

Signed-off-by: Graham whaley <graham.whaley@intel.com>