check: Do not run all linters by default
Some linters were turned off for the CI runs, but not for
a default `make`. This meant a simple `make` would fail the
static checks.

Add a new env var `FULL_STATIC_CHECKS`, and only run all the
linters if this is non-zero. This lets us run the same set
of static checks for a basic `make`, but allows developers to
easily over-ride and run all the static checks if they need.

Fixes: #646

Signed-off-by: Graham whaley <graham.whaley@intel.com>