utils: Make script install components locally
Rather than installing some parts we need into /tmp,
install them into a tree-local directory.

This is cleaner, and when you reboot your machine your
install files will not get cleaned away.

As the same time, only invoke sudo for the parts of the
script we really need to, which keeps the file perms of
the local parts more correct, and accessible by the user.

Signed-off-by: Graham whaley <graham.whaley@intel.com>