virtcontainers: Switch the repository to read-only mode
This commit moves the original README to README-original.md in order
to keep it, and modify the new README to indicate the new location of
virtcontainers in github.com/kata-containers/runtime.

This also introduces a new file error_package_moved.go which prevents
the package from building properly. This is done on purpose, to avoid
confusion for the users of this package, forcing them to update their
dependency to the right repository.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>